feat: simplified deno dockerfile
