+++ _deno/README.md
@@ -1,0 +1,7 @@
+```
+deno run dev
+```
+
+```
+open http://localhost:5173
+```
